lint and rename
